Rename validation module & function
